data-enrichment/2-entity-linking/pyproject.toml
[tool.poetry]
name = "xcurator-entity-linking"
version = "0.0.0" # placeholder | version is defined in git tags | command: poetry version
description = "Extracting Entities in texts and link the them on wikidata, wikipedia and gnd"
authors = ["sraeuchle <sraeuchle@3pc.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
3pc-py-config = {path="./lib/3pc-py-config-1.1.1.tar.gz"}
radboud-el="0.0.1"
spacy="3.7.2"
torch = { version = "2.1.1", source="pytorch-cu118" }
torchvision = { version = "0.16.1", source="pytorch-cu118" }
boto3="1.28.75"
transformers="4.34.1"
jupyter = "^1.0.0"
wikimapper = "0.2.0"
flair = "0.13.0"
pandas= "2.1.3"
openpyxl="3.1.2"

#######################################
#   COMMON Settings
#######################################
[tool.poetry.group.dev.dependencies]
pre-commit = "^2.6"
pytest = "7.4.2"
pytest-asyncio = "0.21.0"

[tool.poetry.build]
generate-setup-file = false

[[tool.poetry.source]]
name = 'pypi-dreipc'
url = 'https://nexus.3pc.de/repository/pypi-group/simple/'
priority = 'supplemental'

[[tool.poetry.source]]
name = 'pytorch-cu118'
url = 'https://download.pytorch.org/whl/cu118'
priority = 'supplemental'

# https://pypi.org/project/poetry-dynamic-versioning/
[tool.poetry-dynamic-versioning]
enable = false
vcs = 'git'
style = "semver"
pattern = '(?x)^v*((?P<epoch>\d+)!)?(?P<base>\d+(\.\d+)*)' # 0.0.0 || v0.0.0 (git tag)

[build-system]
requires = ["poetry-core", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend" # wrapper around poetry.core.masonry.api

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests"
]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"